# support truncate partitioned table and drop partitioned table