--- a/ddl/db_test.go
+++ b/ddl/db_test.go
@@ -2707,10 +2707,10 @@ func (s *testDBSuite) TestTruncatePartitionAndDropTable(c *C) {
 	for i := 0; i < 100; i++ {
 		s.mustExec(c, "insert into t1 values (?)", i)
 	}
-	result := s.tk.MustQuery("select * from t1;")
+	result := s.tk.MustQuery("select count(*) from t1;")
 	c.Assert(len(result.Rows()), Equals, 100)
 	s.tk.MustExec("truncate table t1;")
-	result = s.tk.MustQuery("select * from t1")
+	result = s.tk.MustQuery("select count(*) from t1")
 	c.Assert(len(result.Rows()), Equals, 0)
 
 	// Test drop common table.
@@ -2719,7 +2719,7 @@ func (s *testDBSuite) TestTruncatePartitionAndDropTable(c *C) {
 	for i := 0; i < 100; i++ {
 		s.mustExec(c, "insert into t2 values (?)", i)
 	}
-	result = s.tk.MustQuery("select * from t2;")
+	result = s.tk.MustQuery("select count(*) from t2;")
 	c.Assert(len(result.Rows()), Equals, 100)
 	s.tk.MustExec("drop table t2;")
 	s.testErrorCode(c, "select * from t2;", tmysql.ErrNoSuchTable)
@@ -2750,7 +2750,7 @@ func (s *testDBSuite) TestTruncatePartitionAndDropTable(c *C) {
 	(8, 'aquarium', '1992-08-04'),
 	(9, 'study desk', '2006-09-16'),
 	(10, 'lava lamp', '1998-12-25');`)
-	result = s.tk.MustQuery("select * from t3;")
+	result = s.tk.MustQuery("select count(*) from t3;")
 	c.Assert(len(result.Rows()), Equals, 10)
 	ctx := s.tk.Se.(sessionctx.Context)
 	is := domain.GetDomain(ctx).InfoSchema()
@@ -2766,7 +2766,7 @@ func (s *testDBSuite) TestTruncatePartitionAndDropTable(c *C) {
 	// Test drop table partition.
 	s.tk.MustExec("drop table if exists t4;")
 	s.tk.MustExec("set @@session.tidb_enable_table_partition=1;")
-	s.tk.MustExec(` create table t4(
+	s.tk.MustExec(`create table t4(
 		id int, name varchar(50), 
 		purchased date
 	)
@@ -2789,7 +2789,7 @@ func (s *testDBSuite) TestTruncatePartitionAndDropTable(c *C) {
 	(8, 'aquarium', '1992-08-04'),
 	(9, 'study desk', '2006-09-16'),
 	(10, 'lava lamp', '1998-12-25');`)
-	result = s.tk.MustQuery("select * from t4;")
+	result = s.tk.MustQuery("select count(*) from t4;")
 	c.Assert(len(result.Rows()), Equals, 10)
 	is = domain.GetDomain(ctx).InfoSchema()
 	oldTblInfo, err = is.TableByName(model.NewCIStr("test"), model.NewCIStr("t4"))