--- a/ddl/ddl.go
+++ b/ddl/ddl.go
@@ -63,7 +63,8 @@ var (
 	EnableSplitTableRegion = false
 
 	// PartitionCountLimit is limit of the number of partitions in a table.
-	// Mysql maximum number of partitions is 8192, reference linking https://dev.mysql.com/doc/refman/5.7/en/partitioning-limitations.html
+	// Mysql maximum number of partitions is 8192, our maximum number of partitions is.
+	// Reference linking https://dev.mysql.com/doc/refman/5.7/en/partitioning-limitations.html.
 	PartitionCountLimit = 1024
 )