--- a/ddl/table.go
+++ b/ddl/table.go
@@ -220,12 +220,7 @@ func onTruncateTable(d *ddlCtx, t *meta.Meta, job *model.Job) (ver int64, _ erro
 		oldPartitionIDs = getPartitionIDs(tblInfo)
 		for _, def := range tblInfo.Partition.Definitions {
 			var pid int64
-			genID := func(txn kv.Transaction) error {
-				var err error
-				pid, err = meta.NewMeta(txn).GenGlobalID()
-				return errors.Trace(err)
-			}
-			err = kv.RunInNewTxn(d.store, true /* retryable */, genID)
+			pid, err := d.genGlobalID()
 			if err != nil {
 				job.State = model.JobStateCancelled
 				return ver, errors.Trace(err)