<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

## What have you changed? (mandatory)
support truncate partitioned table
* the partition table deletes the partition data, but retains the `partitionInfo`.
* each partition reassigns the `partitionID`.

support drop partitioned table
* delete partition data is added.
```sql
tidb> set @@session.tidb_enable_table_partition=1;
Query OK, 0 rows affected (0.00 sec)

tidb> create table tr (id int, name varchar(50), purchased date)
    ->     partition by range( year(purchased) ) (
    ->     partition p0 values less than (1990),
    ->     partition p1 values less than (1995),
    ->     partition p2 values less than (2000),
    ->     partition p3 values less than (2005),
    ->     partition p4 values less than (2010),
    ->     partition p5 values less than (2015)
    ->  );
Query OK, 0 rows affected (0.01 sec)

tidb> insert into tr values
    ->    (1, 'desk organiser', '2003-10-15'),
    ->    (2, 'alarm clock', '1997-11-05'),
    ->    (3, 'chair', '2009-03-10'),
    ->    (4, 'bookcase', '1989-01-10'),
    ->    (5, 'exercise bike', '2014-05-09'),
    ->    (6, 'sofa', '1987-06-05'),
    ->    (7, 'espresso maker', '2011-11-22'),
    ->    (8, 'aquarium', '1992-08-04'),
    ->    (9, 'study desk', '2006-09-16'),
    ->    (10, 'lava lamp', '1998-12-25');
Query OK, 10 rows affected (0.12 sec)
```
master branch
```sql
tidb> truncate table tr;
Query OK, 0 rows affected (0.01 sec)

tidb> select * from tr;
+------+----------------+------------+
| id   | name           | purchased  |
+------+----------------+------------+
|    8 | aquarium       | 1992-08-04 |
|    3 | chair          | 2009-03-10 |
|    9 | study desk     | 2006-09-16 |
|    2 | alarm clock    | 1997-11-05 |
|   10 | lava lamp      | 1998-12-25 |
|    1 | desk organiser | 2003-10-15 |
|    5 | exercise bike  | 2014-05-09 |
|    7 | espresso maker | 2011-11-22 |
|    4 | bookcase       | 1989-01-10 |
|    6 | sofa           | 1987-06-05 |
+------+----------------+------------+
10 rows in set (0.00 sec)
```

truncate-table-partition branch
```sql
tidb> truncate table tr;
Query OK, 0 rows affected (0.01 sec)

tidb> select * from tr;
Empty set (0.00 sec)
```
<!--
Please explain **IN DETAIL** what the changes are in this PR and why they are needed:
- Summarize your change (mandatory)
- How does this PR work? Need a brief introduction for the changed logic (optional)
- Separately describe each logical change and avoid lazy messages (optional)
- Describe any limitations of the current code (optional)

Please **NOTE** that:
- Do not assume reviewers understand the original issue
-->

## What is the type of the changes? (mandatory)

<!--
The currently defined types are listed below, please pick one of the types for this PR by removing the others.
-->
New feature
- New feature (non-breaking change which adds functionality)
- Improvement (non-breaking change which is an improvement to an existing feature)
- Bug fix (non-breaking change which fixes an issue)
- Breaking change (fix or feature that would cause existing functionality to not work as expected)

## How has this PR been tested? (mandatory)
unit test

<!--
Please describe the tests that you ran to verify your changes.
Have you finished unit tests, integration tests, or manual tests?
What additional tests would give you greater confidence in this change?
-->

## Does this PR affect documentation (docs/docs-cn) update? (mandatory)
No
<!--
If there is document change, please file a PR in ([docs](https://github.com/pingcap/docs) or [docs-cn](https://github.com/pingcap/docs-cn)) and add the PR number here.
-->

## Does this PR affect tidb-ansible update? (mandatory)
No
<!--
If there is a configuration or metrics change, please file a PR in [tidb-ansible](https://github.com/pingcap/tidb-ansible), and add the PR number here.
-->

## Does this PR need to be added to the release notes? (mandatory)
Yes
<!--
If this PR needs to be added to the release notes, please:
1. add the "**release-note**" label for this PR
2. write your release note in the below block
3. if the PR requires additional action from users switching to the new
   release, describe the concrete action that users need to take in detail.

An example:
```
release note:
// put your release notes for this PR here.

action required:
// put your required action in detail here.
```
-->


## Refer to a related PR or issue link (optional)

## Benchmark result if necessary (optional)

## Add a few positive/negative examples (optional)